requirements.txt
colorama==0.4.4
cffi==1.14.0
construct==2.8.8
cryptg==0.2.post0
ipaddress==1.0.16
Petrovich==1.0.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.19
PySocks==1.7.1
rsa==4.0
six==1.14.0
Telethon==1.10.10
transitions==0.4.1
